enhance organizing of files
