+++ src/Pages/Auth/AuthOperations/Auth.css
@@ -5,7 +5,7 @@
       white 50%,
       transparent 50%
     ),
-    url("../../images/form-bg.jpg") !important;
+    url("../../../images/form-bg.jpg") !important;
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
@@ -69,7 +69,7 @@
         rgba(255, 255, 255, 0.8) 0%,
         rgba(255, 255, 255, 0.8) 100%
       ),
-      url("../../images/form-bg.jpg");
+      url("../../../images/form-bg.jpg");
   }
 
   .custom-form {
