+++ src/Pages/Auth/AuthOperations/Login.js
@@ -1,33 +1,36 @@
 import axios from "axios";
 import React, { useEffect, useRef, useState } from "react";
-import { BaseURL, REGISTER } from "../../Api/Api";
-import Loading from "../../Components/Loading/Loading";
+import { BaseURL, LOGIN } from "../../../Api/Api";
+import Loading from "../../../Components/Loading/Loading";
 import Cookie from "cookie-universal";
-import google from "../../images/googl.png";
+import "../../../Css/Components/googl.css";
+import google from "../../../images/googl.png";
 import Form from "react-bootstrap/Form";
 import { useNavigate } from "react-router-dom";
 
-export default function Regiser() {
+export default function Login() {
   // States for the form
   const [form, setForm] = useState({
-    name: "",
     email: "",
     password: "",
   });
-  const navigate= useNavigate();
+  const navigate = useNavigate();
 
   // Cookies
   const cookie = Cookie();
+
   // State for error
   const [error, setError] = useState("");
 
-    // ref
-    const focus = useRef();
+  // ref
+  const focus = useRef();
 
-    // handle focus
-    useEffect(() => {
-      focus.current.focus();
-    })
+  // handle focus
+  useEffect(() => {
+    focus.current.focus();
+  })
+
+
 
   // loading state
   const [loading, setLoading] = useState(false);
@@ -43,18 +46,20 @@ export default function Regiser() {
     setLoading(true);
     console.log(form);
     try {
-      await axios.post(`${BaseURL}/${REGISTER}`, form).then((res) => {
+      await axios.post(`${BaseURL}/${LOGIN}`, form).then((res) => {
         console.log(res);
         setLoading(false);
         const token = res.data.token;
+        const role = res.data.user.role;
+        const go = role === "1995" ? "users" : "writer";
         cookie.set("e-commerce", token);
-        navigate("/", { replace: true });
+        window.location.pathname = `/dashboard/${go}`;
       });
     } catch (err) {
       console.log(err);
       setLoading(false);
-      if (err.response.status === 422) {
-        setError("Email already been taken");
+      if (err.response.status === 401) {
+        setError("Wrong Email or Password");
       } else {
         setError("Internal Server Error");
       }
@@ -70,33 +75,17 @@ export default function Regiser() {
             height: "100vh",
           }}
         >
-          <Form className="form " onSubmit={handleFormSubmit}>
-            <h1 className="mb-4">Register Now</h1>
+          <Form className="form" onSubmit={handleFormSubmit}>
+            <h1 className="my-4">Login Now</h1>
             <div className="custom-form">
               <Form.Group
                 className="form-custom"
                 controlId="exampleForm.ControlInput1"
               >
                 <Form.Control
-                  type="text"
-                  name="name"
-                  ref={focus}
-                  className="form-control"
-                  placeholder="Enter your Name.."
-                  value={form.name}
-                  onChange={handleFormChange}
-                  required
-                />
-                <Form.Label>Name </Form.Label>
-              </Form.Group>
-
-              <Form.Group
-                className="form-custom"
-                controlId="exampleForm.ControlInput2"
-              >
-                <Form.Control
                   type="email"
                   name="email"
+                  ref={focus}
                   className="form-control"
                   placeholder="Enter your Email.."
                   value={form.email}
@@ -108,7 +97,7 @@ export default function Regiser() {
 
               <Form.Group
                 className="form-custom"
-                controlId="exampleForm.ControlInput3"
+                controlId="exampleForm.ControlInput2  "
               >
                 <Form.Control
                   type="password"
@@ -125,7 +114,7 @@ export default function Regiser() {
               </Form.Group>
 
               <button type="submit" className="btn btn-success">
-                Register
+                Login
               </button>
 
               <div className="google-btn">
