+++ src/Pages/Dashboard/User/User.js
@@ -1,25 +1,53 @@
 import React, { useEffect, useRef, useState } from "react";
 import { Form } from "react-bootstrap";
-import { Axios } from "../../Api/axios";
-import { USER } from "../../Api/Api";
-import Loading from "../../Components/Loading/Loading";
+import { Axios } from "../../../Api/axios";
+import { USER } from "../../../Api/Api";
+import Loading from "../../../Components/Loading/Loading";
+import { useNavigate, useParams } from "react-router-dom";
 
-export default function AddUser() {
+export default function User() {
   const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [role, setRole] = useState("");
-  const [password, setPassword] = useState("");
+  const [disabled, setDisabled] = useState(true);
   const [loading, setLoading] = useState(false);
+  const nav = useNavigate();
+
+  const {id} = useParams()
+  console.log(id);
+
+  const text = useRef("text")
+  console.log(text);
+
+  //   get user data
+
+  useEffect(() => {
+    setLoading(true);
+    Axios.get(`${USER}/${id}`)
+      .then((res) => {
+        console.log(res.data);
+        setName(res.data.name);
+        setEmail(res.data.email);
+        setRole(res.data.role);
+        setLoading(false);
+      })
+      .catch((err) => {
+        console.log(err);
+        nav("/dashboard/users/page/404", { replace: true });
+      })
+      .finally(() => {
+        setDisabled(false);
+      });
+  }, []);
 
   //   function handleSubmit
   async function handleSubmit(e) {
     setLoading(true);
     e.preventDefault();
     try {
-      const res = await Axios.post(`${USER}/add`, {
+      const res = await Axios.post(`${USER}/edit/${id}`, {
         name,
         email,
-        password,
         role,
       });
       console.log(res);
@@ -28,14 +56,6 @@ export default function AddUser() {
       console.log(err);
     }
   }
-
-    // ref
-    const focus = useRef();
-
-    // handle focus
-    useEffect(() => {
-      focus.current.focus();
-    })
 
   //   console.log(name);
   //   console.log(email);
@@ -49,7 +69,6 @@ export default function AddUser() {
           <Form.Label>User Name</Form.Label>
           <Form.Control
             type="text"
-            ref={focus}
             placeholder="name..."
             value={name}
             onChange={(e) => setName(() => e.target.value)}
@@ -67,17 +86,6 @@ export default function AddUser() {
         </Form.Group>
 
         <Form.Group className="mb-3" controlId="exampleForm.ControlInput3">
-          <Form.Label>Password</Form.Label>
-          <Form.Control
-            type="password"
-            placeholder="Enter password.."
-            required
-            value={password}
-            onChange={(e) => setPassword(() => e.target.value)}
-          />
-        </Form.Group>
-
-        <Form.Group className="mb-3" controlId="exampleForm.ControlInput4">
           <Form.Label>Role</Form.Label>
           <Form.Select
             value={role}
@@ -93,14 +101,7 @@ export default function AddUser() {
           </Form.Select>
         </Form.Group>
 
-        <button
-          className="btn btn-primary"
-          disabled={
-            name === "" || email === "" || role === "" || password === ""
-              ? true
-              : false
-          }
-        >
+        <button className="btn btn-primary" disabled={disabled}>
           Save Edits
         </button>
       </Form>
