+++ src/index.js
@@ -5,7 +5,7 @@ import "./Css/Components/Button.css";
 import "./Css/Components/alerts.css";
 import "./Css/Components/Loading.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import "./Pages/Auth/Auth.css";
+import "./Pages/Auth/AuthOperations/Auth.css";
 
 import App from "./App";
 import { BrowserRouter as Router } from "react-router-dom";
